Move the base AthenaProxy class to the proxies module.
